requirements.txt
# Web Scraping (替代 GitHub API)
beautifulsoup4==4.12.3
lxml==5.1.0

# OpenAI API
openai==1.12.0

# Database ORM
SQLAlchemy==2.0.25
alembic==1.13.1

# Web Framework
fastapi==0.109.2
uvicorn==0.27.1

# Environment Variables
python-dotenv==1.0.1

# Data Processing
pandas==2.2.0
tabulate==0.9.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Scheduling
schedule==1.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
